refactor: 변수 명, 클래스 명 변경
